Moved webstore API client code to tool/webstore_client.dart
